# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory Auspac Documentation",
        "description": "Project Inventory Auspac Laravel 12 Documentation",
        "contact": {
            "email": "example@email.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/suppliers": {
            "get": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Get all suppliers",
                "description": "Return a list of all suppliers.",
                "operationId": "dae515d720ee812c126ef52b0b2a82c9",
                "responses": {
                    "200": {
                        "description": "Supplier List.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Supplier"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Unexpected server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Create a new supplier",
                "description": "Create a new supplier.",
                "operationId": "4f02944bfbebc8c3a33209a42ec43695",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Supplier"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Supplier created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Supplier"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/api/suppliers/{id}": {
            "get": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Get a specific supplier",
                "description": "Return a specific supplier by its ID.",
                "operationId": "920374b5be0fda63be39c5c50106e86b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the supplier",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supplier found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Supplier"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Supplier not found"
                    },
                    "500": {
                        "description": "Unexpected server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Update a supplier",
                "description": "Update an existing supplier by its ID.",
                "operationId": "c2b81c027b12fa451256697c8846b3b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the supplier",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Supplier"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Supplier updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Supplier"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Supplier not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "500": {
                        "description": "Unexpected server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Delete a supplier",
                "description": "Delete an existing supplier by its ID.",
                "operationId": "2cf025c2a5b1ee0d4e46daa9478b2866",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the supplier",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supplier deleted successfully"
                    },
                    "404": {
                        "description": "Supplier not found"
                    },
                    "500": {
                        "description": "Unexpected server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Supplier": {
                "title": "Supplier",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Supplier 1"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}